Done UT, controller, exception handler

+++ src/main/java/com/example/shopdemoitsj/service/impl/ItemServiceImpl.java
@@ -1,15 +1,22 @@
 package com.example.shopdemoitsj.service.impl;
 
 import com.example.shopdemoitsj.dto.ItemDto;
+import com.example.shopdemoitsj.exception.ItemCascadeDeleteError;
 import com.example.shopdemoitsj.exception.ItemNotFoundException;
 import com.example.shopdemoitsj.mapper.ItemMapper;
+import com.example.shopdemoitsj.model.CartDetail;
 import com.example.shopdemoitsj.model.Item;
+import com.example.shopdemoitsj.model.OrderDetail;
+import com.example.shopdemoitsj.repository.CartDetailRepository;
 import com.example.shopdemoitsj.repository.ItemRepository;
+import com.example.shopdemoitsj.repository.OrderDetailRepository;
 import com.example.shopdemoitsj.service.ItemService;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,6 +25,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(rollbackFor = Throwable.class)
 public class ItemServiceImpl implements ItemService {
   @Autowired ItemRepository itemRepository;
+
+  @Autowired CartDetailRepository cartDetailRepository;
+
+  @Autowired OrderDetailRepository orderDetailRepository;
 
   @Override
   public List<ItemDto> findAll() {
@@ -46,7 +57,18 @@ public class ItemServiceImpl implements ItemService {
   }
 
   @Override
-  public void delete(int itemId) {
-    itemRepository.deleteById(itemId);
+  public ResponseEntity<HttpStatus> delete(int itemId) throws ItemCascadeDeleteError, Exception {
+    List<CartDetail> cartDetailList = cartDetailRepository.findByItemId(itemId);
+    List<OrderDetail> orderDetailList = orderDetailRepository.findByItemId(itemId);
+
+    if (!cartDetailList.isEmpty() || !orderDetailList.isEmpty()) {
+      throw new ItemCascadeDeleteError();
+    }
+    try {
+      itemRepository.deleteById(itemId);
+    } catch (Exception ex) {
+      throw new Exception();
+    }
+    return new ResponseEntity<>(HttpStatus.OK);
   }
 }
